requirements.txt
blinker==1.9.0
bleach==6.1.0  # HTML sanitization (>=5.0 for newsletter feature)
cachelib==0.13.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
Flask==2.3.3
Flask-Login==0.6.3
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.33.0
msgspec==0.19.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.5
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==2.3.7
